Create postBuild
wget is a program which retrieves content from web servers. This line extracts the content from the bitly URL and saves it to the filename denoted by the -O flag (capital “O”, not zero), in this case gapminder.csv. The -q flag tells wget to do this quietly, meaning it won’t print anything to the console.